all cards displaying again
