+++ src/game/components.rs
@@ -1,6 +1,7 @@
 use std::cmp::Ordering;
 use bevy::prelude::*;
-use super::{CardLocation, CardVal, Suit};
+use rand::{seq::SliceRandom, SeedableRng};
+use rand_xoshiro::Xoshiro256PlusPlus;
 
 #[derive(Component, Debug, Clone, Copy, PartialEq, Eq)]
 pub struct  Card {
@@ -102,7 +103,6 @@ pub struct Pile;
 #[derive(Debug, Clone, Component)]
 pub struct PlayerTurnText;
 
-
 #[derive(Debug, Clone, Component)]
 pub struct DeadCards {
     cards: Vec<Card>,
@@ -114,4 +114,105 @@ pub struct Player {
     pub facedown_cards: Vec<Card>,
     pub faceup_cards: Vec<Card>,
     pub hand: Vec<Card>,
+}
+
+#[derive(Debug, Resource)]
+pub struct LocalPlayers(pub Vec<u64>);
+
+#[derive(States, Default, Debug, PartialEq, Eq, Clone, Hash)]
+pub enum DeckState {
+    #[default]
+    Ordered,
+    Shuffled,
+    Dealt,
+    Gameplay,
+    Display,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum CardVal {
+    Two = 0,
+    Three = 1,
+    Four = 2,
+    Five = 3,
+    Six = 4,
+    Seven = 5,
+    Eight = 6,
+    Nine = 7,
+    Ten = 8,
+    Jack = 9,
+    Queen = 10,
+    King = 12,
+    Ace = 13,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
+pub enum Suit {
+    Club,
+    Heart,
+    Spade,
+    Diamond,
+}
+
+#[derive(Clone, Copy, Debug)]
+pub enum CardLocation {
+    LPHand,
+    LPFaceUp,
+    RPHand,
+    RPFaceUp,
+    Deck,
+    Dead,
+    Pile,
+}
+
+#[derive(Resource, Debug, Clone)]
+pub struct CardDeck{
+    pub cards: CardGroup,
+}
+
+impl Default for CardDeck {
+    fn default() -> Self {
+        // ordered deck of cards
+        let mut deck = Vec::new();
+
+        for &suit in &[Suit::Heart, Suit::Diamond, Suit::Club, Suit::Spade] {
+            for &number in &[
+                CardVal::Two, CardVal::Three, CardVal::Four, CardVal::Five,
+                CardVal::Six, CardVal::Seven, CardVal::Eight, CardVal::Nine,
+                CardVal::Ten, CardVal::Jack, CardVal::Queen, CardVal::King, CardVal::Ace
+            ] {
+                deck.push(Card { number, suit });
+            }
+        }
+
+        Self {
+            cards: CardGroup {
+                cards: deck,
+                location: CardLocation::Deck,
+            }
+        }
+    }
+}
+
+impl CardDeck {
+    pub fn shuffle(&mut self, seed: u64) {
+        // let mut rng = thread_rng().next_u64();
+        let mut rng = Xoshiro256PlusPlus::seed_from_u64(seed);
+        self.cards.cards.shuffle(&mut rng);
+    }
+}
+
+#[derive(Resource, Default, Debug, PartialEq, Eq, Clone, Hash)]
+pub struct PlayerTurn{
+    pub ids: Vec<u64>,
+    pub turn: usize,
+}
+impl PlayerTurn {
+    // only for 2 player atm
+    pub fn next(&mut self) {
+        self.turn += 1;
+        if self.turn >= self.ids.len() {
+            self.turn = 0;
+        }
+    }
 }
